requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-crispy-forms==2.3
django-formtools==2.5.1
pillow==11.0.0
python-decouple==3.8
sqlparse==0.5.2
tzdata==2024.2
